Move PubkeyListMatch rule test to separate file
